+++ packages/varlet-cli/varlet.default.config.js
@@ -30,7 +30,7 @@ module.exports = {
         'en-US': 'English',
       },
       github: 'https://github.com/varletjs/varlet',
-      playground: 'https://varlet.gitee.io/varlet-ui-playground',
+      playground: 'https://varlet-ui-playground.vercel.app',
       darkMode: true,
     },
     clipboard: {