+++ packages/varlet-icons/varlet-icons.config.js
@@ -2,5 +2,5 @@ module.exports = {
   namespace: 'var-icon',
   fileName: 'varlet-icons',
   fontName: 'varlet-icons',
-  publicPath: 'https://cdn.jsdelivr.net/npm/@varlet/icons/dist/fonts/',
+  base64: true,
 }