+++ packages/varlet-ui/src/tabs/Tabs.vue
@@ -8,7 +8,8 @@
           n(`--item-${itemDirection}`),
           n(`--layout-${layoutDirection}-padding`),
           [elevation, 'var-elevation--4'],
-          [fixedBottom, n('--fixed-bottom')]
+          [fixedBottom, n('--fixed-bottom')],
+          [safeArea, n('--safe-area')]
         )
       "
       :style="{ background: color }"
@@ -86,8 +87,8 @@ export default defineComponent({
       return tabList.find(({ name }: TabProvider) => props.active === name.value)
     }
 
-    const matchIndex = (): TabProvider | undefined => {
-      return tabList.find(({ index }: TabProvider) => props.active === index.value)
+    const matchIndex = (activeIndex?: number): TabProvider | undefined => {
+      return tabList.find(({ index }: TabProvider) => (activeIndex ?? props.active) === index.value)
     }
 
     const matchBoundary = (): TabProvider | undefined => {
@@ -97,13 +98,11 @@ export default defineComponent({
 
       const { active } = props
 
-      isNumber(active)
-        ? active > length.value - 1
-          ? call(props['onUpdate:active'], length.value - 1)
-          : call(props['onUpdate:active'], 0)
-        : null
-
-      return matchIndex()
+      if (isNumber(active)) {
+        const activeIndex = active > length.value - 1 ? length.value - 1 : 0
+        call(props['onUpdate:active'], activeIndex)
+        return matchIndex(activeIndex)
+      }
     }
 
     const watchScrollable = () => {